PAZHOU/base/requirements.txt
paddlepaddle-gpu==2.3.2
fvcore==0.1.5.post20220119
torchvision==0.8.2
cloudpickle==2.0.0
omegaconf==2.1.1
pycocotools==2.0.4
timm==0.4.12
typing_extensions==4.5.0
paddleseg==2.7.0
paddledet==2.6.0
opencv-python==4.4.0.46
hydra-core==1.1.1
sacred==0.8.2
tensorboard==2.5.0
